fix(store): get_deposit_count reads i32 not i64 — panic at startup (e2e)
The Cantina #9 LET-divergence monitor's `get_deposit_count` impl read
service_state.deposit_counter as i64, but migrations/001_initial.sql
declares the column as `INT NOT NULL` (postgres int4 / Rust i32).
tokio-postgres panics with "error retrieving column 0: error
deserializing column 0" when the type doesn't match.

The panic surfaced on the FIRST sync tick after startup because the
LET-divergence monitor runs every on_post_sync. Symptom in the e2e
stack: proxy panicked → MidenClient receiver closed → next
eth_sendRawTransaction failed with "MidenClient::with: failed to
queue a request - receiver is closed" → /health reported "node
connection lost".

Fixed by reading as i32 (matches the schema). Cast to u64 for the
public API is unchanged.

Caught running scripts/e2e-security.sh against the docker-compose
stack — 40/41 passed, the one fail being /health: degraded.

+++ src/store/postgres.rs
@@ -1024,7 +1024,9 @@ impl Store for PgStore {
         let row = client
             .query_one("SELECT deposit_counter FROM service_state WHERE id = 1", &[])
             .await?;
-        let val: i64 = row.get(0);
+        // service_state.deposit_counter is `INT NOT NULL` (postgres int4 / Rust i32),
+        // not BIGINT. Reading as i64 panics with "error deserializing column 0".
+        let val: i32 = row.get(0);
         Ok(val as u64)
     }
